python/openapi-core/requirements.txt
openapi-core
# pylint
# mypy
PyYAML
